requirements.txt
numpy
pandas
unstructured
tiktoken
pypdf
pinecone-client
openai
langchain
python-dotenv
faiss-cpu